+++ src/solenoids/on_off_solenoid.h
@@ -12,7 +12,7 @@ public:
     void off();
     bool is_on();
 private:
-    uint64_t on_time_ms;
+    uint32_t on_time_ms;
     uint32_t target_on_time;
     uint16_t target_hold_pwm;
     bool state = false;
